Sucessfully called OpenWeatherAPI
